requirements.txt
-i https://pypi.org/simple
click==7.1.2
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-login==0.5.0
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
pyjwt==1.7.1
python-dotenv==0.14.0
six==1.15.0
sqlalchemy==1.3.19
werkzeug==1.0.1
wtforms==2.3.3
flask-praetorian==1.0.0
flask-session==0.3.2
boto3 ==1.16.15
